更新内容: - [重要] 支持了 `SCA` 检测 - [SCA] 支持 `Apache Log4j2` 全版本的匹配 - [SCA] 支持 `FAT JAR` 和 `LIB JAR` 两种检测

+++ src/main/java/me/n1ar4/jar/analyzer/sca/SCAUtil.java
@@ -1,50 +1,73 @@
 package me.n1ar4.jar.analyzer.sca;
 
-import me.n1ar4.log.LogManager;
-import me.n1ar4.log.Logger;
+import me.n1ar4.jar.analyzer.utils.IOUtil;
 
 import java.io.*;
+import java.util.ArrayList;
 import java.util.Enumeration;
+import java.util.List;
 import java.util.jar.JarEntry;
 import java.util.jar.JarFile;
-import java.util.jar.JarInputStream;
 
 public class SCAUtil {
-    private static final Logger logger = LogManager.getLogger();
+    public static List<String> visitedJar = new ArrayList<>();
+    public static boolean found = false;
+    public static byte[] data = null;
+
+    public static void refresh() {
+        visitedJar.clear();
+        found = false;
+        data = null;
+    }
 
     @SuppressWarnings("all")
     public static byte[] exploreJar(File file, String keyClassName) {
+        if (found) {
+            return data;
+        }
         try (JarFile jarFile = new JarFile(file)) {
             Enumeration<JarEntry> entries = jarFile.entries();
             while (entries.hasMoreElements()) {
                 JarEntry entry = entries.nextElement();
                 if (entry.getName().endsWith(".class")) {
                     if (entry.getName().contains(keyClassName)) {
-                        return getClassBytes(jarFile, entry);
+                        found = true;
+                        data = getClassBytes(jarFile, entry);
+                        return data;
                     }
                 } else if (entry.getName().endsWith(".jar")) {
+                    if (visitedJar.contains(entry.getName())) {
+                        continue;
+                    }
                     File nestedJarFile = extractNestedJar(jarFile, entry);
                     exploreJar(nestedJarFile, keyClassName);
+                    visitedJar.add(entry.getName());
                     nestedJarFile.delete();
                 }
             }
-        } catch (IOException e) {
-            logger.error(e.toString());
-            SCALogger.logger.error(e.toString()); // 如果需要，也可以保留这行
+        } catch (IOException ignored) {
         }
-        return null;
+        return data;
     }
 
     @SuppressWarnings("all")
     private static File extractNestedJar(JarFile jarFile, JarEntry entry) throws IOException {
+        if (jarFile == null) {
+            throw new IllegalArgumentException("JarFile cannot be null");
+        }
+        if (entry == null) {
+            throw new IllegalArgumentException("JarEntry cannot be null");
+        }
+        if (jarFile.getJarEntry(entry.getName()) == null) {
+            throw new IOException("JarEntry does not exist in the JarFile");
+        }
         File tempFile = File.createTempFile("nested-", ".jar");
-        try (JarInputStream jarInputStream = new JarInputStream(jarFile.getInputStream(entry));
-             FileInputStream fis = new FileInputStream(tempFile)) {
-            int bytesRead;
-            byte[] buffer = new byte[1024];
-            while ((bytesRead = jarInputStream.read(buffer)) != -1) {
-                fis.read(buffer, 0, bytesRead);
+        try (InputStream jarInputStream = jarFile.getInputStream(entry);
+             FileOutputStream fos = new FileOutputStream(tempFile)) {
+            if (jarInputStream == null) {
+                throw new IOException("Could not get InputStream for JarEntry");
             }
+            IOUtil.copy(jarInputStream, fos);
         }
         return tempFile;
     }
